+++ CHANGELOG.md
@@ -8,4 +8,6 @@
 
 - Threads with Image
 - Threads with Link Attachment
+- Get Post Id From URL
+- Reply to Other Threads
 - Fixed major issues